+++ lib/td_dq/rules/rule_implementation.ex
@@ -25,11 +25,46 @@ defmodule TdDq.Rules.RuleImplementation do
       :system_params,
       :rule_id
     ])
-    |> validate_required([
+    |> validate_required(required_attrs(rule_implementation))
+    |> validate_length(:implementation_key, max: 255)
+  end
+
+  defp required_attrs(%RuleImplementation{
+         rule:
+           %Rule{
+             rule_type:
+               %TdDq.Rules.RuleType{params: %{"system_params" => system_params}} = rule_type
+           } = rule
+       }) do
+    system_required =
+      system_params
+      |> Enum.find(fn param -> Map.get(param, "name") == "system_required" end)
+      |> case do
+        nil -> true
+        system_required_props -> Map.get(system_required_props, "value")
+      end
+
+    case system_required do
+      true ->
+        [
+          :system,
+          :system_params,
+          :rule_id
+        ]
+
+      false ->
+        [
+          :system_params,
+          :rule_id
+        ]
+    end
+  end
+
+  defp required_attrs(_any) do
+    [
       :system,
       :system_params,
       :rule_id
-    ])
-    |> validate_length(:implementation_key, max: 255)
+    ]
   end
 end