+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [Unreleased]
+
+## Added
+
+- [TD-2048] New Rule Type system param ´system_required´ of type boolean with hidden property
+
+## Changed
+
+- [TD-2048] Attribute system in Rule Implementation requirement is defined in Rule Type system params as ´system_required´
+
 ## [3.7.0] 2019-09-30
 
 ### Added